requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
yt-dlp>=2025.3.31
certifi>=2025.1.31
PyNaCl>=1.5.0
openai>=1.0.0
